reservation_card_service/pyproject.toml
[project]
name = "reservation-card-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
